requirements.txt
disnake
colorama
mysql-connector-python
python-dotenv
mctools